fix some types

+++ src/ast.hpp
@@ -26,7 +26,7 @@ namespace core {
 
 	class ASTNode : public std::enable_shared_from_this<ASTNode>, public CodePosition {
 	public:
-		ASTNode(unsigned int row, unsigned int col)
+		ASTNode(size_t row, size_t col)
 			: CodePosition(row, col) {
 		}
 
@@ -35,7 +35,7 @@ namespace core {
 
 	class ASTStatementNode : public ASTNode {
 	public:
-		ASTStatementNode(unsigned int row, unsigned int col)
+		ASTStatementNode(size_t row, size_t col)
 			: ASTNode(row, col) {
 		}
 
@@ -44,12 +44,12 @@ namespace core {
 
 	class ASTExprNode : public ASTNode {
 	public:
-		ASTExprNode(unsigned int row, unsigned int col)
+		ASTExprNode(size_t row, size_t col)
 			: ASTNode(row, col) {
 		}
 
 		void accept(Visitor*) override = 0;
-		virtual long long hash(Visitor*) = 0;
+		virtual intmax_t hash(Visitor*) = 0;
 	};
 
 	class ASTProgramNode : public ASTNode {
@@ -70,7 +70,7 @@ namespace core {
 		std::string identifier;
 
 		ASTBuiltinCallNode(std::string identifier,
-			unsigned int row, unsigned int col);
+			size_t row, size_t col);
 
 		void accept(Visitor*) override;
 	};
@@ -80,7 +80,7 @@ namespace core {
 		std::vector<std::string> library;
 
 		ASTUsingNode(const std::vector<std::string>& library,
-			unsigned int row, unsigned int col);
+			size_t row, size_t col);
 
 		void accept(Visitor*) override;
 	};
@@ -89,21 +89,21 @@ namespace core {
 	public:
 		std::string name_space;
 
-		ASTNamespaceManagerNode(const std::string& name_space, unsigned int col, unsigned int row);
+		ASTNamespaceManagerNode(const std::string& name_space, size_t col, size_t row);
 
 		void accept(Visitor*) override = 0;
 	};
 
 	class ASTIncludeNamespaceNode : public ASTNamespaceManagerNode {
 	public:
-		ASTIncludeNamespaceNode(const std::string& name_space, unsigned int col, unsigned int row);
+		ASTIncludeNamespaceNode(const std::string& name_space, size_t col, size_t row);
 
 		void accept(Visitor*) override;
 	};
 
 	class ASTExcludeNamespaceNode : public ASTNamespaceManagerNode {
 	public:
-		ASTExcludeNamespaceNode(const std::string& name_space, unsigned int col, unsigned int row);
+		ASTExcludeNamespaceNode(const std::string& name_space, size_t col, size_t row);
 
 		void accept(Visitor*) override;
 	};
@@ -117,7 +117,7 @@ namespace core {
 		ASTDeclarationNode(const std::string& identifier, Type type, Type array_type,
 			const std::vector<std::shared_ptr<ASTExprNode>>& dim, const std::string& type_name,
 			const std::string& type_name_space, std::shared_ptr<ASTExprNode> expr, bool is_const,
-			unsigned int row, unsigned int col);
+			size_t row, size_t col);
 
 		void accept(Visitor*) override;
 	};
@@ -130,7 +130,7 @@ namespace core {
 		ASTUnpackedDeclarationNode(Type type, Type array_type, const std::vector<std::shared_ptr<ASTExprNode>>& dim,
 			const std::string& type_name, const std::string& type_name_space,
 			const std::vector<std::shared_ptr<ASTDeclarationNode>>& declarations, std::shared_ptr<ASTExprNode> expr,
-			unsigned int row, unsigned int col);
+			size_t row, size_t col);
 
 		void accept(Visitor*) override;
 	};
@@ -144,7 +144,7 @@ namespace core {
 		std::shared_ptr<ASTExprNode> expr;
 
 		ASTAssignmentNode(const std::vector<Identifier>& identifier_vector, const std::string& name_space,
-			const std::string& op, std::shared_ptr<ASTExprNode> expr, unsigned int row, unsigned int col);
+			const std::string& op, std::shared_ptr<ASTExprNode> expr, size_t row, size_t col);
 
 		void accept(Visitor*) override;
 	};
@@ -153,7 +153,7 @@ namespace core {
 	public:
 		std::shared_ptr<ASTExprNode> expr;
 
-		ASTReturnNode(std::shared_ptr<ASTExprNode> expr, unsigned int row, unsigned int col);
+		ASTReturnNode(std::shared_ptr<ASTExprNode> expr, size_t row, size_t col);
 
 		void accept(Visitor*) override;
 	};
@@ -162,21 +162,21 @@ namespace core {
 	public:
 		std::vector<std::shared_ptr<ASTNode>> statements;
 
-		ASTBlockNode(const std::vector<std::shared_ptr<ASTNode>>& statements, unsigned int row, unsigned int col);
+		ASTBlockNode(const std::vector<std::shared_ptr<ASTNode>>& statements, size_t row, size_t col);
 
 		void accept(Visitor*) override;
 	};
 
 	class ASTContinueNode : public ASTStatementNode {
 	public:
-		ASTContinueNode(unsigned int row, unsigned int col);
+		ASTContinueNode(size_t row, size_t col);
 
 		void accept(Visitor*) override;
 	};
 
 	class ASTBreakNode : public ASTStatementNode {
 	public:
-		ASTBreakNode(unsigned int row, unsigned int col);
+		ASTBreakNode(size_t row, size_t col);
 
 		void accept(Visitor*) override;
 	};
@@ -185,7 +185,7 @@ namespace core {
 	public:
 		std::shared_ptr<ASTExprNode> exit_code;
 
-		ASTExitNode(std::shared_ptr<ASTExprNode> exit_code, unsigned int row, unsigned int col);
+		ASTExitNode(std::shared_ptr<ASTExprNode> exit_code, size_t row, size_t col);
 
 		void accept(Visitor*) override;
 	};
@@ -193,14 +193,14 @@ namespace core {
 	class ASTSwitchNode : public ASTStatementNode {
 	public:
 		std::shared_ptr<ASTExprNode> condition;
-		std::map<std::shared_ptr<ASTExprNode>, unsigned int> case_blocks;
-		std::map<unsigned int, unsigned int> parsed_case_blocks;
-		unsigned int default_block;
+		std::map<std::shared_ptr<ASTExprNode>, size_t> case_blocks;
+		std::map<size_t, size_t> parsed_case_blocks;
+		size_t default_block;
 		std::vector<std::shared_ptr<ASTNode>> statements;
 
 		ASTSwitchNode(std::shared_ptr<ASTExprNode> condition, const std::vector<std::shared_ptr<ASTNode>>& statements,
-			const std::map<std::shared_ptr<ASTExprNode>, unsigned int>& case_blocks,
-			unsigned int default_block, unsigned int row, unsigned int col);
+			const std::map<std::shared_ptr<ASTExprNode>, size_t>& case_blocks,
+			size_t default_block, size_t row, size_t col);
 
 		void accept(Visitor*) override;
 	};
@@ -213,7 +213,7 @@ namespace core {
 		std::shared_ptr<ASTBlockNode> else_block;
 
 		ASTIfNode(std::shared_ptr<ASTExprNode> condition, std::shared_ptr<ASTBlockNode> if_block, const std::vector<std::shared_ptr<ASTElseIfNode>>& else_ifs,
-			std::shared_ptr<ASTBlockNode> else_block, unsigned int row, unsigned int col);
+			std::shared_ptr<ASTBlockNode> else_block, size_t row, size_t col);
 
 		void accept(Visitor*) override;
 	};
@@ -223,7 +223,7 @@ namespace core {
 		std::shared_ptr<ASTExprNode> condition;
 		std::shared_ptr<ASTBlockNode> block;
 
-		ASTElseIfNode(std::shared_ptr<ASTExprNode> condition, std::shared_ptr<ASTBlockNode> block, unsigned int row, unsigned int col);
+		ASTElseIfNode(std::shared_ptr<ASTExprNode> condition, std::shared_ptr<ASTBlockNode> block, size_t row, size_t col);
 
 		void accept(Visitor*) override;
 	};
@@ -232,7 +232,7 @@ namespace core {
 	public:
 		std::vector<std::string> identifiers;
 
-		ASTEnumNode(const std::vector<std::string>& identifiers, unsigned int row, unsigned int col);
+		ASTEnumNode(const std::vector<std::string>& identifiers, size_t row, size_t col);
 
 		void accept(Visitor*) override;
 	};
@@ -243,7 +243,7 @@ namespace core {
 		std::shared_ptr<ASTBlockNode> try_block;
 		std::shared_ptr<ASTBlockNode> catch_block;
 
-		ASTTryCatchNode(std::shared_ptr<ASTStatementNode> decl, std::shared_ptr<ASTBlockNode> try_block, std::shared_ptr<ASTBlockNode> catch_block, unsigned int row, unsigned int col);
+		ASTTryCatchNode(std::shared_ptr<ASTStatementNode> decl, std::shared_ptr<ASTBlockNode> try_block, std::shared_ptr<ASTBlockNode> catch_block, size_t row, size_t col);
 
 		void accept(Visitor*) override;
 	};
@@ -252,14 +252,14 @@ namespace core {
 	public:
 		std::shared_ptr<ASTExprNode> error;
 
-		ASTThrowNode(std::shared_ptr<ASTExprNode> error, unsigned int row, unsigned int col);
+		ASTThrowNode(std::shared_ptr<ASTExprNode> error, size_t row, size_t col);
 
 		void accept(Visitor*) override;
 	};
 
 	class ASTEllipsisNode : public ASTStatementNode {
 	public:
-		ASTEllipsisNode(unsigned int row, unsigned int col);
+		ASTEllipsisNode(size_t row, size_t col);
 
 		void accept(Visitor*) override;
 	};
@@ -269,7 +269,7 @@ namespace core {
 		std::array<std::shared_ptr<ASTNode>, 3> expressions;
 		std::shared_ptr<ASTBlockNode> block;
 
-		ASTForNode(const std::array<std::shared_ptr<ASTNode>, 3>& expressions, std::shared_ptr<ASTBlockNode> block, unsigned int row, unsigned int col);
+		ASTForNode(const std::array<std::shared_ptr<ASTNode>, 3>& expressions, std::shared_ptr<ASTBlockNode> block, size_t row, size_t col);
 
 		void accept(Visitor*) override;
 	};
@@ -280,7 +280,7 @@ namespace core {
 		std::shared_ptr<ASTNode> collection;
 		std::shared_ptr<ASTBlockNode> block;
 
-		ASTForEachNode(std::shared_ptr<ASTNode> itdecl, std::shared_ptr<ASTNode> collection, std::shared_ptr<ASTBlockNode> block, unsigned int row, unsigned int col);
+		ASTForEachNode(std::shared_ptr<ASTNode> itdecl, std::shared_ptr<ASTNode> collection, std::shared_ptr<ASTBlockNode> block, size_t row, size_t col);
 
 		void accept(Visitor*) override;
 	};
@@ -290,14 +290,14 @@ namespace core {
 		std::shared_ptr<ASTExprNode> condition;
 		std::shared_ptr<ASTBlockNode> block;
 
-		ASTWhileNode(std::shared_ptr<ASTExprNode> condition, std::shared_ptr<ASTBlockNode> block, unsigned int row, unsigned int col);
+		ASTWhileNode(std::shared_ptr<ASTExprNode> condition, std::shared_ptr<ASTBlockNode> block, size_t row, size_t col);
 
 		void accept(Visitor*) override;
 	};
 
 	class ASTDoWhileNode : public ASTWhileNode {
 	public:
-		ASTDoWhileNode(std::shared_ptr<ASTExprNode> condition, std::shared_ptr<ASTBlockNode> block, unsigned int row, unsigned int col);
+		ASTDoWhileNode(std::shared_ptr<ASTExprNode> condition, std::shared_ptr<ASTBlockNode> block, size_t row, size_t col);
 
 		void accept(Visitor*) override;
 	};
@@ -308,7 +308,7 @@ namespace core {
 		std::map<std::string, VariableDefinition> variables;
 
 		ASTStructDefinitionNode(const std::string& identifier, const std::map<std::string, VariableDefinition>& variables,
-			unsigned int row, unsigned int col);
+			size_t row, size_t col);
 
 		void accept(Visitor*) override;
 	};
@@ -320,8 +320,8 @@ namespace core {
 		std::shared_ptr<ASTBlockNode> block;
 
 		ASTFunctionDefinitionNode(const std::string& identifier, const std::vector<TypeDefinition*>& parameters,
-			Type type, const std::string& type_name, const std::string& type_name_space, Type array_type, const std::vector<unsigned int>& dim,
-			std::shared_ptr<ASTBlockNode> block, unsigned int row, unsigned int col);
+			Type type, const std::string& type_name, const std::string& type_name_space, Type array_type, const std::vector<size_t>& dim,
+			std::shared_ptr<ASTBlockNode> block, size_t row, size_t col);
 
 		void accept(Visitor*) override;
 	};
@@ -331,40 +331,40 @@ namespace core {
 	public:
 		T value;
 
-		ASTLiteralNode(T value, unsigned int row, unsigned int col) : ASTExprNode(row, col), value(value) {};
+		ASTLiteralNode(T value, size_t row, size_t col) : ASTExprNode(row, col), value(value) {};
 
 		void accept(Visitor*) override;
-		virtual long long hash(Visitor*) override;
+		virtual intmax_t hash(Visitor*) override;
 	};
 
 	class ASTValueNode : public ASTExprNode {
 	public:
 		Value* value;
 
-		ASTValueNode(Value* value, unsigned int row, unsigned int col);
+		ASTValueNode(Value* value, size_t row, size_t col);
 
 		void accept(Visitor*) override;
-		virtual long long hash(Visitor*) override;
+		virtual intmax_t hash(Visitor*) override;
 	};
 
 	class ASTLambdaFunction : public ASTExprNode {
 	public:
 		std::shared_ptr<ASTFunctionDefinitionNode> fun;
 
-		ASTLambdaFunction(std::shared_ptr<ASTFunctionDefinitionNode> fun, unsigned int row, unsigned int col);
+		ASTLambdaFunction(std::shared_ptr<ASTFunctionDefinitionNode> fun, size_t row, size_t col);
 
 		void accept(Visitor*) override;
-		virtual long long hash(Visitor*) override;
+		virtual intmax_t hash(Visitor*) override;
 	};
 
 	class ASTArrayConstructorNode : public ASTExprNode {
 	public:
 		std::vector<std::shared_ptr<ASTExprNode>> values;
 
-		ASTArrayConstructorNode(const std::vector<std::shared_ptr<ASTExprNode>>& values, unsigned int row, unsigned int col);
+		ASTArrayConstructorNode(const std::vector<std::shared_ptr<ASTExprNode>>& values, size_t row, size_t col);
 
 		void accept(Visitor*) override;
-		virtual long long hash(Visitor*) override;
+		virtual intmax_t hash(Visitor*) override;
 	};
 
 	class ASTStructConstructorNode : public ASTExprNode {
@@ -374,26 +374,26 @@ namespace core {
 		std::map<std::string, std::shared_ptr<ASTExprNode>> values;
 
 		ASTStructConstructorNode(const std::string& type_name, const std::string& name_space,
-			const std::map<std::string, std::shared_ptr<ASTExprNode>>& values, unsigned int row, unsigned int col);
+			const std::map<std::string, std::shared_ptr<ASTExprNode>>& values, size_t row, size_t col);
 
 		void accept(Visitor*) override;
-		virtual long long hash(Visitor*) override;
+		virtual intmax_t hash(Visitor*) override;
 	};
 
 	class ASTNullNode : public ASTExprNode {
 	public:
-		ASTNullNode(unsigned int row, unsigned int col);
+		ASTNullNode(size_t row, size_t col);
 
 		void accept(Visitor*) override;
-		virtual long long hash(Visitor*) override;
+		virtual intmax_t hash(Visitor*) override;
 	};
 
 	class ASTThisNode : public ASTExprNode {
 	public:
-		ASTThisNode(unsigned int row, unsigned int col);
+		ASTThisNode(size_t row, size_t col);
 
 		void accept(Visitor*) override;
-		virtual long long hash(Visitor*) override;
+		virtual intmax_t hash(Visitor*) override;
 	};
 
 	class ASTBinaryExprNode : public ASTExprNode {
@@ -402,10 +402,10 @@ namespace core {
 		std::shared_ptr<ASTExprNode> left;
 		std::shared_ptr<ASTExprNode> right;
 
-		ASTBinaryExprNode(const std::string& op, std::shared_ptr<ASTExprNode> left, std::shared_ptr<ASTExprNode> right, unsigned int row, unsigned int col);
+		ASTBinaryExprNode(const std::string& op, std::shared_ptr<ASTExprNode> left, std::shared_ptr<ASTExprNode> right, size_t row, size_t col);
 
 		void accept(Visitor*) override;
-		virtual long long hash(Visitor*) override;
+		virtual intmax_t hash(Visitor*) override;
 	};
 
 	class ASTUnaryExprNode : public ASTExprNode {
@@ -413,10 +413,10 @@ namespace core {
 		std::string unary_op;
 		std::shared_ptr<ASTExprNode> expr;
 
-		ASTUnaryExprNode(const std::string& unary_op, std::shared_ptr<ASTExprNode> expr, unsigned int row, unsigned int col);
+		ASTUnaryExprNode(const std::string& unary_op, std::shared_ptr<ASTExprNode> expr, size_t row, size_t col);
 
 		void accept(Visitor*) override;
-		virtual long long hash(Visitor*) override;
+		virtual intmax_t hash(Visitor*) override;
 	};
 
 	class ASTIdentifierNode : public ASTExprNode {
@@ -425,10 +425,10 @@ namespace core {
 		std::string name_space;
 		std::vector<Identifier> identifier_vector;
 
-		explicit ASTIdentifierNode(const std::vector<Identifier>& identifier_vector, std::string name_space, unsigned int row, unsigned int col);
+		explicit ASTIdentifierNode(const std::vector<Identifier>& identifier_vector, std::string name_space, size_t row, size_t col);
 
 		void accept(Visitor*) override;
-		virtual long long hash(Visitor*) override;
+		virtual intmax_t hash(Visitor*) override;
 	};
 
 	class ASTTernaryNode : public ASTExprNode {
@@ -438,10 +438,10 @@ namespace core {
 		std::shared_ptr<ASTExprNode> value_if_false;
 
 		ASTTernaryNode(std::shared_ptr<ASTExprNode> condition, std::shared_ptr<ASTExprNode> value_if_true,
-			std::shared_ptr<ASTExprNode> value_if_false, unsigned int row, unsigned int col);
+			std::shared_ptr<ASTExprNode> value_if_false, size_t row, size_t col);
 
 		void accept(Visitor*) override;
-		virtual long long hash(Visitor*) override;
+		virtual intmax_t hash(Visitor*) override;
 	};
 
 	class ASTInNode : public ASTExprNode {
@@ -449,10 +449,10 @@ namespace core {
 		std::shared_ptr<ASTExprNode> value;
 		std::shared_ptr<ASTExprNode> collection;
 
-		ASTInNode(std::shared_ptr<ASTExprNode> value, std::shared_ptr<ASTExprNode> collection, unsigned int row, unsigned int col);
+		ASTInNode(std::shared_ptr<ASTExprNode> value, std::shared_ptr<ASTExprNode> collection, size_t row, size_t col);
 
 		void accept(Visitor*) override;
-		virtual long long hash(Visitor*) override;
+		virtual intmax_t hash(Visitor*) override;
 	};
 
 	class ASTFunctionCallNode : public ASTExprNode {
@@ -465,10 +465,10 @@ namespace core {
 		std::shared_ptr<ASTFunctionCallNode> expression_call;
 
 		ASTFunctionCallNode(const std::string& name_space, const std::vector<Identifier>& identifier_vector, const std::vector<std::shared_ptr<ASTExprNode>>& parameters,
-			std::vector<Identifier> expression_identifier_vector, std::shared_ptr<ASTFunctionCallNode> expression_call, unsigned int row, unsigned int col);
+			std::vector<Identifier> expression_identifier_vector, std::shared_ptr<ASTFunctionCallNode> expression_call, size_t row, size_t col);
 
 		void accept(Visitor*) override;
-		virtual long long hash(Visitor*) override;
+		virtual intmax_t hash(Visitor*) override;
 	};
 
 	class ASTTypeCastNode : public ASTExprNode {
@@ -476,54 +476,54 @@ namespace core {
 		Type type;
 		std::shared_ptr<ASTExprNode> expr;
 
-		ASTTypeCastNode(Type type, std::shared_ptr<ASTExprNode> expr, unsigned int row, unsigned int col);
+		ASTTypeCastNode(Type type, std::shared_ptr<ASTExprNode> expr, size_t row, size_t col);
 
 		void accept(Visitor*) override;
-		virtual long long hash(Visitor*) override;
+		virtual intmax_t hash(Visitor*) override;
 	};
 
 	class ASTTypeNode : public ASTExprNode {
 	public:
 		TypeDefinition type;
 
-		ASTTypeNode(TypeDefinition type, unsigned int row, unsigned int col);
+		ASTTypeNode(TypeDefinition type, size_t row, size_t col);
 
 		void accept(Visitor*) override;
-		virtual long long hash(Visitor*) override;
+		virtual intmax_t hash(Visitor*) override;
 	};
 
 	class ASTCallOperatorNode : public ASTExprNode {
 	public:
 		std::shared_ptr<ASTExprNode> expr;
 
-		ASTCallOperatorNode(std::shared_ptr<ASTExprNode> expr, unsigned int row, unsigned int col);
+		ASTCallOperatorNode(std::shared_ptr<ASTExprNode> expr, size_t row, size_t col);
 
 		void accept(Visitor*) override = 0;
-		virtual long long hash(Visitor*) override = 0;
+		virtual intmax_t hash(Visitor*) override = 0;
 	};
 
 	class ASTTypeOfNode : public ASTCallOperatorNode {
 	public:
-		ASTTypeOfNode(std::shared_ptr<ASTExprNode> expr, unsigned int row, unsigned int col);
+		ASTTypeOfNode(std::shared_ptr<ASTExprNode> expr, size_t row, size_t col);
 
 		void accept(Visitor*) override;
-		virtual long long hash(Visitor*) override;
+		virtual intmax_t hash(Visitor*) override;
 	};
 
 	class ASTTypeIdNode : public ASTCallOperatorNode {
 	public:
-		ASTTypeIdNode(std::shared_ptr<ASTExprNode> expr, unsigned int row, unsigned int col);
+		ASTTypeIdNode(std::shared_ptr<ASTExprNode> expr, size_t row, size_t col);
 
 		void accept(Visitor*) override;
-		virtual long long hash(Visitor*) override;
+		virtual intmax_t hash(Visitor*) override;
 	};
 
 	class ASTRefIdNode : public ASTCallOperatorNode {
 	public:
-		ASTRefIdNode(std::shared_ptr<ASTExprNode> expr, unsigned int row, unsigned int col);
+		ASTRefIdNode(std::shared_ptr<ASTExprNode> expr, size_t row, size_t col);
 
 		void accept(Visitor*) override;
-		virtual long long hash(Visitor*) override;
+		virtual intmax_t hash(Visitor*) override;
 	};
 
 }
