+++ src/md_builtin.cpp
@@ -226,7 +226,7 @@ void ModuleBuiltin::build_decls() {
 		std::make_shared<ASTBuiltinCallNode>(Constants::BUILTIN_NAMES.at(BuintinFuncs::READCH), 0, 0)}, 0, 0)));
 
 	parameters = std::vector<TypeDefinition*>();
-	variable = new VariableDefinition("it", Type::T_ANY, std::vector<unsigned int>());
+	variable = new VariableDefinition("it", Type::T_ANY, std::vector<size_t>());
 	parameters.push_back(variable);
 	func_decls.emplace(Constants::BUILTIN_NAMES.at(BuintinFuncs::LEN) + "A", FunctionDefinition(Constants::BUILTIN_NAMES.at(BuintinFuncs::LEN) + "A", Type::T_INT, parameters,
 		std::make_shared<ASTBlockNode>(std::vector<std::shared_ptr<ASTNode>>{
